pcd: mount the corruption accessors under proof
